pyproject.toml
[project]
name = "artefact"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "minio>=7.2.20",
    "pandas>=3.0.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]
